[update] ディレクトリを変更, コンポーネント名を変更
